마커 controller, js

+++ back/Giboo/src/main/java/kh/fin/giboo/volunteer/model/vo/Volunteer.java
@@ -13,8 +13,8 @@ import java.sql.Date;
 public class Volunteer {
     private int volunteerNo;
     private int targetPeople;
-    private int locationX;
-    private int locationY;
+    private double locationX;
+    private double locationY;
     private String volunteerAddr;
     private String volunteerTime;
     private String enrollDate;
